requirements.txt
streamlit-antd-components
streamlit-pdf-viewer
firebase-admin
streamlit-free-text-select
extra_streamlit_components
pyjwt
google-auth-oauthlib
google-api-python-client
streamlit-extras